feat: Main functionalities

+++ src/main/java/com/restfulapi/desafio/services/PlanoService.java
@@ -9,8 +9,12 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import com.restfulapi.desafio.dtos.PlanoDto;
+import com.restfulapi.desafio.exceptions.CodigoCadastroAnsInvalid;
+import com.restfulapi.desafio.exceptions.PlanoAtrelado;
+import com.restfulapi.desafio.exceptions.PlanoInvalid;
 import com.restfulapi.desafio.exceptions.PlanoNotFound;
 import com.restfulapi.desafio.model.Plano;
+import com.restfulapi.desafio.repositories.BeneficiarioRepository;
 import com.restfulapi.desafio.repositories.PlanoRepository;
 
 @Service
@@ -19,33 +23,46 @@ public class PlanoService {
     @Autowired
     PlanoRepository planoRepository;
 
+    @Autowired
+    BeneficiarioRepository beneficiarioRepository;
+
     public List<Plano> getAll(){
         return planoRepository.findAll();
     }
 
     public Plano getById(UUID id) {
-    return planoRepository.findById(id)
-            .orElseThrow(() -> new PlanoNotFound("Plano com ID " + id + " não encontrado."));
+    return planoRepository.findById(id).orElseThrow(() -> new PlanoNotFound());
     }
 
 
     public Plano save(PlanoDto dto){
         var plano = new Plano();
-        BeanUtils.copyProperties(dto, plano);
+        // String codigoRegistroAns = dto.codigo_registro_ans();
 
-                
+        
+        BeanUtils.copyProperties(dto, plano);
         return planoRepository.save(plano);    
         
     }
 
     public void delete(UUID id){
-        Plano plano = planoRepository.findById(id).orElseThrow(() -> new RuntimeException("Plano não encontrado"));
+        Plano plano = planoRepository.findById(id).orElseThrow(() -> new PlanoNotFound());
+        boolean beneficiarioExistente = beneficiarioRepository.existsByPlanoId(id);
+
+        if(!planoRepository.existsById(id)){
+            throw new PlanoInvalid();
+        }
+
+        if(beneficiarioExistente){
+            throw new PlanoAtrelado();
+        }
+
         planoRepository.delete(plano);
     }
 
     public Plano update(UUID id, PlanoDto dto){
 
-        Plano plano = planoRepository.findById(id).orElseThrow(() -> new RuntimeException("Fudeu"));
+        Plano plano = planoRepository.findById(id).orElseThrow(() -> new PlanoNotFound());
 
         BeanUtils.copyProperties(dto, plano, "id", "data_cadastro");
